Add panning methods to NodifyEditor
